backend/skills' System/skill_normalization_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
torch>=1.13.0
transformers>=4.30.0
pytest==7.4.3
httpx==0.25.2
